Implement phase 1.7 (confirmed) + 1.8: collection/set-browse/portfolio API
1.7 (collection, collection_item, sealed_holding, goal, goal_item
models + migration) turns out to already be fully done -- those models
and the baseline migration landed in the earlier session. No new model
code needed; noted here so it's not mistaken for a gap.

1.8 adds the service + API layer:

- app/services/prices.py: reads the current_price view (bindparam
  expanding IN-list), converting SQLite's raw float return through
  Decimal(str(...)) rather than straight from float -- the same
  never-float-for-money rule CLAUDE.md states.
- app/services/sets.py: list_sets + get_set_detail, returning the
  set's cards with each variant's current market price and owned
  quantity (PRD T2/T3), and per-card owned/needed flags computed from
  whether a *canonical* variant has quantity > 0.
- app/services/collection.py: collection_item CRUD, upserting on the
  documented natural key (collection_id, card_variant_id, condition,
  language, is_graded, grade) so re-adding the same holding updates
  quantity in place rather than duplicating a row.
- app/services/portfolio.py: total market value, cost basis, and
  unrealized gain (PRD T6) over a collection's holdings.

All four return plain dataclasses, not ORM objects or Pydantic models,
so they stay reusable from a future CLI command without pulling in
FastAPI -- per the layer rules in docs/01-architecture.md. The API
layer (app/api/schemas.py, app/api/routers/) is a thin Pydantic/FastAPI
shell over these, wired into main.py under /api/v1.

Also fixes a real, pre-existing bug this surfaced: every money column
(price_point.low/mid/high/market/direct_low, sealed_product.msrp,
collection_item/sealed_holding.acquired_price) was typed
Mapped[float | None] despite being Numeric(10,2) columns that
SQLAlchemy actually returns as Decimal at runtime -- directly
contradicting CLAUDE.md's "Money is Numeric(10,2) in the DB and
Decimal in Python. Never float for currency." Fixed the type hints to
Decimal (no schema change -- confirmed with `alembic check`, still
no-op) and fixed the one call site that was assigning a raw float
into a now-Decimal column (ingest/prices.py's price_point upsert).

Tests: service-level tests seed the DB directly and assert on the
returned dataclasses; test_api.py smoke-tests every route via
FastAPI's TestClient with a dependency-overridden in-memory DB
(StaticPool -- SQLite's plain ":memory:" DSN gives each thread a
different empty database, which TestClient's threaded dispatch hits
immediately).

+++ backend/app/models/catalog.py
@@ -1,5 +1,6 @@
 """Card catalogue: sets, cards, printing variants, sealed products, prices."""
 from datetime import date
+from decimal import Decimal
 
 from sqlalchemy import (
     JSON,
@@ -97,7 +98,7 @@ class SealedProduct(Base, TimestampMixin):
     pack_config_id: Mapped[int | None] = mapped_column(
         ForeignKey("pull_rate_profile.id", ondelete="SET NULL")
     )
-    msrp: Mapped[float | None] = mapped_column(Numeric(10, 2))
+    msrp: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
     contains_promos: Mapped[bool] = mapped_column(Boolean, default=False)
     image_url: Mapped[str | None] = mapped_column(String(512))
 
@@ -122,9 +123,9 @@ class PricePoint(Base):
     sub_type_name: Mapped[str] = mapped_column(String(64))
     observed_on: Mapped[date] = mapped_column(Date)
     source: Mapped[str] = mapped_column(String(32), default="tcgcsv")
-    low: Mapped[float | None] = mapped_column(Numeric(10, 2))
-    mid: Mapped[float | None] = mapped_column(Numeric(10, 2))
-    high: Mapped[float | None] = mapped_column(Numeric(10, 2))
-    market: Mapped[float | None] = mapped_column(Numeric(10, 2))
-    direct_low: Mapped[float | None] = mapped_column(Numeric(10, 2))
+    low: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
+    mid: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
+    high: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
+    market: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
+    direct_low: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
     currency: Mapped[str] = mapped_column(String(3), default="USD")
